Move StatsProvider to correct package
