+++ src/main/java/space/npstr/wolfia/domain/stats/BotStatsCommand.java
@@ -21,7 +21,6 @@ import space.npstr.wolfia.commands.BaseCommand;
 import space.npstr.wolfia.commands.CommandContext;
 import space.npstr.wolfia.commands.PublicCommand;
 import space.npstr.wolfia.domain.Command;
-import space.npstr.wolfia.utils.StatsProvider;
 
 import javax.annotation.Nonnull;
